refactor(js): fold the single-file events/ domain into core
events/ held one constants file; every consumer already treats it as
foundation. It now lives at core/event-names with the same export
surface from the root barrel.
